+++ CHANGELOG.md
@@ -71,3 +71,4 @@ believed to be representative.
 - Fixed `opacity` attribute handling (use `srcOver`)
 - Support `mix-blend-mode` attribute
 - Implement clip-rule attribute
+- Implement text-decoration attribute